Build: build a release tarball that can be loaded in to eccube
We create a `release/` directory that we can (later) commit to a
separate branch so that we can provide a direct download link of the
tarball from github.

I am not clear on what the contents of the tarball itself are supposed
to be, so I used exactly what Akira put in his tarball in commit ad9446b.
This still needs to be tested.

We also drop our top-level `README.md` in there so it appears if anybody
happens to look at that branch in GitHub. We may want to replace this
with a different README more suited to that particular branch, however.